// src/pages/user/utility/enum/user.enum.ts
export enum genderEnum {
  Male = 1,
  Female = 2,
  Other = 3,
}

export enum MonthsEnum {
  January = 1,
  February = 2,
  March = 3,
  April = 4,
  May = 5,
  June = 6,
  July = 7,
  August = 8,
  September = 9,
  October = 10,
  November = 11,
  December = 12,
}

export enum CategoryTypeEnum {
  Income = 1,
  Expense = 2,
}

export enum ChartEnum {
  ExpenseCount = 1,
  TotalExpense = 2,
  IncomeCount = 3,
  TotalIncome = 4,
}